Re-updating CryoDRGN plugin in cryodrgn-reform branch
