+++ cryodrgn/protocols/protocol_analyze.py
@@ -2,6 +2,7 @@
 # *
 # * Authors:     Grigory Sharov (gsharov@mrc-lmb.cam.ac.uk) [1]
 # *              Yunior C. Fonseca Reyna (cfonseca@cnb.csic.es) [2]
+# *              Eduardo García Delgado (eduardo.garcia@cnb.csic.es) [2]
 # *
 # * [1] MRC Laboratory of Molecular Biology (MRC-LMB)
 # * [2] Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
@@ -26,38 +27,33 @@
 # *
 # **************************************************************************
 
-import os
+import os, shutil
 import numpy as np
-from enum import Enum
-
+import pickle
+from pyworkflow.protocol.constants import *
 import pyworkflow.utils as pwutils
+from pyworkflow.object import *
 import pyworkflow.protocol.params as params
-from pyworkflow.constants import NEW
-import pyworkflow.object as pwobj
-from pwem.protocols import ProtAnalysis3D
-from pwem.objects import SetOfVolumes, Volume
-from pwem.emlib.image import ImageHandler, DT_FLOAT
+from pyworkflow.constants import PROD
+import pwem.objects as emobj
+from pwem.emlib import *
+from pwem.emlib.image import ImageHandler
+from pwem.protocols import ProtProcessParticles, ProtFlexBase
+from .. import Plugin
+from ..constants import *
 
-from cryodrgn import Plugin
-from cryodrgn.constants import (EPOCH_LAST, EPOCH_SELECTION, Z_VALUES,
-                                AB_INITIO_HOMO, CLUSTER_WARD, V3_3_2)
-from cryodrgn.protocols.protocol_base import CryoDrgnProtBase
-
-
-class outputs(Enum):
-    Volumes = SetOfVolumes
-
-
-class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
+class CryoDrgnProtAnalyze(ProtProcessParticles, ProtFlexBase):
     """ CryoDrgn protocol to visualize latent space and generate volumes. """
 
-    _label = "analyze results"
-    _devStatus = NEW
-    _possibleOutputs = outputs
+    _label = "cryodrgn analyze"
+    _devStatus = PROD
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
 
     def _createFilenameTemplates(self):
         """ Centralize how files are called within the protocol. """
-        out = lambda p: self.getOutputDir(f'analyze.{self._epoch}', p)
+        out = lambda p: self._getOutputDir(f'analyze.{self._epoch}', p)
 
         myDict = {
             'input_mask': self._getExtraPath("input_mask.mrc"),
@@ -76,12 +72,6 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
     # --------------------------- DEFINE param functions ----------------------
     def _defineParams(self, form):
         form.addSection(label='Input')
-        form.addHidden(params.GPU_LIST, params.StringParam, default='0',
-                       label="Choose GPU IDs",
-                       help="GPU may have several cores. Set it to zero"
-                            " if you do not know what we are talking about."
-                            " First core index is 0, second 1 and so on."
-                            " You can use only a single GPU.")
         form.addParam('inputProt', params.PointerParam, important=True,
                       pointerClass='CryoDrgnProtTrain, CryoDrgnProtAbinitio',
                       label="Previous run to analyse")
@@ -98,10 +88,13 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
         form.addSection('Volume generation')
         form.addParam('doFlip', params.BooleanParam, default=False,
                       label="Flip handedness of output volumes")
+
         form.addParam('doInvert', params.BooleanParam, default=False,
                       label="Invert contrast of output volumes")
+
         form.addParam('doDownsample', params.BooleanParam, default=False,
                       label="Downsample volumes?")
+
         form.addParam('boxSize', params.IntParam, default=128,
                       condition='doDownsample', label="New box size (px)")
 
@@ -116,9 +109,11 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
                            "remaining on the data manifold since we don't want "
                            "to generate structures from unoccupied regions of "
                            "the latent space.")
+
         form.addParam('pc', params.IntParam, default=2,
                       label="Number of principal components",
                       help="Number of principal component traversals to generate.")
+
         form.addParam('ksamples', params.IntParam, default=20,
                       label='Number of K-means samples to generate',
                       help="*cryodrgn analyze* uses the k-means clustering "
@@ -143,18 +138,21 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
         form.addParam('numVols', params.IntParam, default=500,
                       condition='doLandscape',
                       label="Number of volumes to generate")
-        group = form.addGroup('Masking',
-                              condition='doLandscape')
+
+        group = form.addGroup('Masking', condition='doLandscape')
         group.addParam('autoMask', params.BooleanParam, default=True,
                        label="Mask volumes automatically?")
+
         group.addParam('inputMask', params.PointerParam, important=True,
                        condition="not autoMask",
                        pointerClass='VolumeMask', allowsNull=True,
                        label="Custom mask")
+
         group.addParam('threshold', params.FloatParam, default=0.,
                        condition="autoMask",
                        label="Threshold for masking",
                        help="Default 0 means a half of max density value")
+
         group.addParam('dilate', params.IntParam, default=5,
                        condition="autoMask",
                        label="Dilation (px)",
@@ -167,8 +165,19 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
                        default=CLUSTER_WARD,
                        display=params.EnumParam.DISPLAY_HLIST,
                        label="Linkage for agglomerative clustering")
+
         group.addParam('numClusters', params.IntParam, default=10,
                        label="Number of clusters")
+
+        form.addHidden(params.GPU_LIST, params.StringParam, default='0',
+                       label="Choose GPU IDs",
+                       help="GPU may have several cores. Set it to zero"
+                            " if you do not know what we are talking about."
+                            " First core index is 0, second 1 and so on."
+                            " You can use multiple GPUs - in that case"
+                            " set to i.e. *0 1 2*.")
+
+        form.addParallelSection(threads=1, mpi=1)
 
     # --------------------------- INSERT steps functions ----------------------
     def _insertAllSteps(self):
@@ -182,13 +191,30 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
 
         self._createFilenameTemplates()
 
-        self._insertFunctionStep(self.runAnalysisStep, self._epoch,
-                                 needsGPU=True)
+        if self.doLandscape and self.hasMultLatentVars():
+            self._insertFunctionStep(self.convertInputStep, self._epoch, needsGPU=False)
+
+        self._insertFunctionStep(self.runAnalysisStep, self._epoch, needsGPU=True)
         self._insertFunctionStep(self.createOutputStep, needsGPU=False)
 
     # --------------------------- STEPS functions -----------------------------
+    def convertInputStep(self, epoch):
+        # Copy analyze.epoch/umap.pkl to landscape.epoch folder
+        pwutils.makePath(self._getOutputDir(f'landscape.{epoch}'))
+        pwutils.copyFile(self._getFileName('umaps'),
+                         self._getOutputDir(f'landscape.{epoch}/umap.pkl'))
+
+        if not self.autoMask:
+            # convert mask to mrc
+            maskFn = self.inputMask.get().getFileName()
+            if pwutils.getExt(maskFn) == ".mrc":
+                pwutils.createLink(maskFn, self._getFileName("input_mask"))
+            else:
+                ih = ImageHandler()
+                ih.convert(maskFn, self._getFileName("input_mask"), DT_FLOAT)
+
     def runAnalysisStep(self, epoch):
-        pwutils.makePath(self.getOutputDir())
+        pwutils.makePath(self._getOutputDir())
         self._runProgram('analyze', self._getAnalyzeArgs(epoch))
 
         if self.doGraphTraversal and self.hasMultLatentVars():
@@ -196,7 +222,6 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
             self._runProgram('eval_vol', self._getEvalArgs())
 
         if self.doLandscape and self.hasMultLatentVars():
-            self.convertInputs(epoch)
             self._runProgram('analyze_landscape', self._getLandscapeArgs(epoch))
 
     def createOutputStep(self):
@@ -206,9 +231,8 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
         files, zValues = self._getVolumes()
         setOfVolumes = self._createVolumeSet(files, zValues, fn, samplingRate)
 
-        self._defineOutputs(**{outputs.Volumes.name: setOfVolumes})
-        self._defineSourceRelation(self._getInputProt()._getInputParticles(pointer=True),
-                                   setOfVolumes)
+        self._defineOutputs(outputVolumes=setOfVolumes)
+        self._defineSourceRelation(self._getInputProt()._getInputParticles(), setOfVolumes)
 
     # --------------------------- INFO functions ------------------------------
     def _summary(self):
@@ -266,26 +290,11 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
         return errors
 
     # --------------------------- UTILS functions -----------------------------
-    def convertInputs(self, epoch):
-        # Copy analyze.epoch/umap.pkl to landscape.epoch folder
-        pwutils.makePath(self.getOutputDir(f'landscape.{epoch}'))
-        pwutils.copyFile(self._getFileName('umaps'),
-                         self.getOutputDir(f'landscape.{epoch}/umap.pkl'))
-
-        if not self.autoMask:
-            # convert mask to mrc
-            maskFn = self.inputMask.get().getFileName()
-            if pwutils.getExt(maskFn) == ".mrc":
-                pwutils.createLink(maskFn, self._getFileName("input_mask"))
-            else:
-                ih = ImageHandler()
-                ih.convert(maskFn, self._getFileName("input_mask"), DT_FLOAT)
-
     def _getAnalyzeArgs(self, epoch):
         args = [
             self._getInputProt()._getExtraPath("output"),
             f"{epoch}",
-            f"-o {self.getOutputDir(f'analyze.{epoch}')}",
+            f"-o {self._getOutputDir(f'analyze.{epoch}')}",
             f"--Apix {self._getSamplingRate()}",
             f"--device {self.gpuList.get()}",
             f"-d {self.boxSize}" if self.doDownsample else "",
@@ -330,7 +339,7 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
         args = [
             self._getInputProt()._getExtraPath("output"),
             f"{epoch}",
-            f"-o {self.getOutputDir(f'landscape.{epoch}')}",
+            f"-o {self._getOutputDir(f'landscape.{epoch}')}",
             f"--Apix {self._getSamplingRate()}",
             f"--device {self.gpuList.get()}",
             "--skip-umap",
@@ -402,7 +411,7 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
         :return: a set of volumes
         """
         pwutils.cleanPath(path)
-        volSet = SetOfVolumes(filename=path)
+        volSet = emobj.SetOfVolumes(filename=path)
         volSet.setSamplingRate(samplingRate)
         volSet.setObjComment("k-means sample volumes")
         volId = 0
@@ -411,9 +420,9 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
             zValues = [[i] for i in zValues]
 
         for volFn in files:
-            vol = Volume()
+            vol = emobj.Volume()
             vol.setFileName(volFn)
-            vector = pwobj.CsvList()
+            vector = CsvList()
             # We assume that each row "i" of z_values corresponds to each
             # volumes with ID "i"
             volZValues = zValues[volId]
@@ -427,11 +436,18 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
 
         return volSet
 
+    def _runProgram(self, program, args):
+        gpus = ','.join(str(i) for i in self.getGpuList())
+        self.runJob(Plugin.getProgram(program, gpus), ' '.join(args))
+
     def _getSamplingRate(self):
         return self._getInputProt()._getInputParticles().getSamplingRate()
 
     def _getBoxSize(self):
         return self._getInputProt()._getInputParticles().getXDim()
+    
+    def _getOutputDir(self, *paths):
+        return self._getExtraPath("output", *paths)
 
     def _getOutputSampling(self):
         if self.doDownsample:
@@ -448,8 +464,14 @@ class CryoDrgnProtAnalyze(ProtAnalysis3D, CryoDrgnProtBase):
         else:
             return inputProt.zDim.get() > 1
 
-    def _getLastEpoch(self):
-        return self._getInputProt()._getLastEpoch()
-
     def _getInputProt(self):
         return self.inputProt.get()
+
+    def _getLastEpoch(self):
+        outDir = self._getInputProt()._getExtraPath("output")
+        files = [file for file in os.listdir(outDir) if file.startswith("weights")]
+        print(os.path.basename(files[0]))
+        if len(files) != 0:
+            epochs = [os.path.basename(file).split('.')[1] for file in files]
+            lastEpoch = max([int(epoch) for epoch in epochs if epoch != "pkl"])
+        return lastEpoch
